Move server and zkp into view model
this saves us lots of operations on rotation
