+++ app/src/main/java/nl/tudelft/cs4160/identitychain/main/MainViewModel.kt
@@ -1,12 +1,28 @@
 package nl.tudelft.cs4160.identitychain.main
 
+import android.app.AlertDialog
 import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.LiveDataReactiveStreams
 import android.arch.lifecycle.MutableLiveData
+import android.content.DialogInterface
+import android.util.Log
+import android.widget.Toast
+import com.zeroknowledgeproof.rangeProof.RangeProofTrustedParty
+import io.grpc.Server
+import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import nl.tudelft.cs4160.identitychain.Util.Key
+import nl.tudelft.cs4160.identitychain.block.TrustChainBlock
+import nl.tudelft.cs4160.identitychain.database.TrustChainDBHelper
+import nl.tudelft.cs4160.identitychain.grpc.ChainServiceServer
+import nl.tudelft.cs4160.identitychain.grpc.asMessage
+import nl.tudelft.cs4160.identitychain.message.ChainService
 import nl.tudelft.cs4160.identitychain.network.PeerItem
 import nl.tudelft.cs4160.identitychain.network.ServiceFactory
+import java.net.NetworkInterface
+import java.security.KeyPair
 
 
 class MainViewModel(app: Application) : AndroidViewModel(app) {
@@ -14,6 +30,16 @@ class MainViewModel(app: Application) : AndroidViewModel(app) {
     val serviceFactory = initializeServiceFactory()
     val allPeers: LiveData<PeerItem> = LiveDataReactiveStreams.fromPublisher(serviceFactory.startPeerDiscovery())
     val peerSelection: LiveData<PeerItem> = selectedPeer
+
+    private lateinit var grpc: Server
+    lateinit private var server: ChainServiceServer
+    private val dbHelper: TrustChainDBHelper = TrustChainDBHelper(app)
+
+    val trustedParty = RangeProofTrustedParty()
+    val zkp = trustedParty.generateProof(30, 18, 100)
+
+    val kp by lazy(this::initKeys)
+
 
     fun select(peer: PeerItem) {
         selectedPeer.value = peer
@@ -25,5 +51,95 @@ class MainViewModel(app: Application) : AndroidViewModel(app) {
         return factory
     }
 
+    init {
+        if (isStartedFirstTime) {
+            val block = TrustChainBlock.createGenesisBlock(kp)
+            dbHelper.insertInDB(block)
+        }
 
+        val (server, grpc) = ChainServiceServer.createServer(kp, 8080, localIPAddress!!, dbHelper, this::attestationPrompt, zkp.second)
+        Log.i(TAG, "created server")
+        this.grpc = grpc
+        this.server = server
+
+    }
+
+    /**
+     * Checks if this is the first time the app is started and returns a boolean value indicating
+     * this state.
+     * @return state - false if the app has been initialized before, true if first time app started
+     */
+    // check if a genesis block is present in database
+    val isStartedFirstTime: Boolean
+        get() {
+            val genesisBlock = dbHelper.getBlock(kp.public.encoded, TrustChainBlock.GENESIS_SEQ)
+
+            return genesisBlock == null
+        }
+
+    private fun initKeys(): KeyPair {
+        val kp = Key.loadKeys(getApplication())
+        if (kp == null) {
+            val kp = Key.createAndSaveKeys(getApplication())
+            Log.i(TAG, "New keys created")
+            return kp
+        }
+        return kp
+    }
+
+    /**
+     * Finds the local IP address of this device, loops trough network interfaces in order to find it.
+     * The address that is not a loopback address is the IP of the device.
+     * @return a string representation of the device's IP address
+     */
+    val localIPAddress: String?
+        get() {
+            try {
+                val netInterfaces = NetworkInterface.getNetworkInterfaces()
+                for (netInt in netInterfaces) {
+                    for (addr in netInt.inetAddresses) {
+                        if (addr.isSiteLocalAddress) {
+                            return addr.hostAddress
+                        }
+                    }
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+
+            return null
+        }
+
+    private fun attestationPrompt(attestation: ChainService.PublicSetupResult): Single<Boolean> {
+        val fuckyiou = Single.create<Boolean> { source ->
+            val message = "attest for some cool dude that his _ is between ${attestation.a} - ${attestation.b}"
+            AlertDialog.Builder(getApplication()).setMessage(message)
+                    .setPositiveButton("yes", object : DialogInterface.OnClickListener {
+                        override fun onClick(dialog: DialogInterface?, which: Int) {
+                            source.onSuccess(true)
+                        }
+                    })
+                    .setNegativeButton("no", object : DialogInterface.OnClickListener {
+                        override fun onClick(dialog: DialogInterface?, which: Int) {
+                            source.onSuccess(false)
+                        }
+                    }).show()
+        }
+        return fuckyiou.subscribeOn(AndroidSchedulers.mainThread())
+    }
+
+    fun createClaim(): Single<ChainService.PeerTrustChainBlock>? {
+        val peeritem = peerSelection.value
+        val asMessage: ChainService.PublicSetupResult = zkp.first.asMessage()
+        val publicPayLoad = asMessage.toByteArray()
+        return peeritem?.let { server.sendBlockToKnownPeer(it, publicPayLoad) }
+    }
+
+    override fun onCleared() {
+        grpc.shutdownNow()
+    }
+
+    companion object {
+        val TAG = "MainViewModel"
+    }
 }
